Corrected 'snow' and changed vars double to float
snow variable should be a surface field, but I was creating it
with respect to sea ice. When sea ice changes to be a four level
field, snow should also have been changed to no longer copy sea
ice. Also made changes s.t. default is 'float' and not 'double'