Time stamp in seconds since 2009-04-06 00:00:00
Commit on 27 April correctly interpolated in days, but A. Dawson's
scripts cannot read in files with 'days' timestamp. Therefore have
converted back to seconds since 2009-04-06 00:00:00

Also identified bug in cascade_etadotdpdeta.ncl: the script calc-
ulates a time derivative, and assumes the time is in seconds.
This was therefore wrong in the previous submission, but is
correct now.