Correction to treatment of missing values
earlier version of ecmwf_function's blend_data woudl set an entire
level to zero if any of the values were missing, expecting this to
only occur at highest levels, above CASCADE data's top

However, can get occasional missing values lower in model

We now check whether a level is missing or if it is occasional
values. If a whole level is missing, treat as before

If only a few values are missing, we use bilinear interpolation in
lat and lon to fill them in.